third_party/blink/web_tests/platform/mac-mac11/virtual/threaded-prefer-compositing/external/wpt/css/cssom-view/smooth-scroll-in-load-event-expected.txt

This is a testharness.js-based test.
[FAIL] Smooth scroll in load event handler
  assert_equals: Final value of scrollLeft expected 800 but got 0
Harness: the test ran to completion.
